pyproject.toml
[project]
name = "watchdog"
version = "0.1.0"
description = "A Telegram bot for managing group chats"
readme = "README.md"
authors = [
    { name = "Crol"}
]
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "aiorun>=2025.1.1",
    "aiosqlite>=0.21.0",
    "dacite>=1.9.2",
    "numpy>=2.3.3",
    "platformdirs>=4.4.0",
    "python-telegram-bot>=22.5",
    "toml>=0.10.2",
    "uvloop>=0.21.0",
]

[project.scripts]
watchdog = "watchdog:run"

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"
